// apps/web/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ArticleOut: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
            /** Slug */
            slug: string;
            /**
             * Status
             * @enum {string}
             */
            status: "draft" | "scheduled" | "published" | "cancelled" | "archived";
            category: components["schemas"]["CategoryOut"] | null;
            lead_image: components["schemas"]["MediaAssetOut"] | null;
            /** Published At */
            published_at: string | null;
            /** Sample */
            sample: boolean;
            /** Reading Minutes */
            reading_minutes: number | null;
            /** Translations */
            translations: {
                [key: string]: components["schemas"]["ArticleTranslationOut"];
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}